Hook up Resting and Maximum heart rates in the settings screen and make the rest of the UI take these values into account.

+++ HeartRevs/UI/HeartView.swift
@@ -14,7 +14,7 @@ struct HeartView: View {
     
     @Binding var flipped: Bool
     @Binding var bpm: Double
-    @Binding var maxHR: Int
+    @Binding var maxHR: Double
     @State var animationAmount: CGFloat = 1.0
     
     private let maxAnimationAmount = CGFloat(1.04)
